# Speed up caching files duration