--- a/audb/core/load.py
+++ b/audb/core/load.py
@@ -169,18 +169,17 @@ def _files_duration(
         files: typing.Sequence[str],
         format: typing.Optional[str],
 ):
-
-    def fix_file(file: str):
-        file = os.path.join(db.root, file)
-        if format is not None:
-            file = audeer.replace_file_extension(file, format)
-        return file
-
     field = define.DEPEND_FIELD_NAMES[define.DependField.DURATION]
     durs = deps._df.loc[files][field]
     durs = durs[durs > 0]
     durs = pd.to_timedelta(durs, unit='s')
-    durs.index = durs.index.map(fix_file)
+    root = db.root + os.path.sep
+    if format is not None:
+        cur_ext = re.compile(r'\.[a-zA-Z0-9]+$')  # match file extension
+        new_ext = f'.{format}'
+        durs.index = root + durs.index.str.sub(cur_ext, new_ext)
+    else:
+        durs.index = root + durs.index
     db._files_duration = durs.to_dict()