--- a/audb/core/load.py
+++ b/audb/core/load.py
@@ -173,11 +173,10 @@ def _files_duration(
     durs = deps._df.loc[files][field]
     durs = durs[durs > 0]
     durs = pd.to_timedelta(durs, unit='s')
+    durs.index.name = 'file'
     if format is not None:
-        cur_ext = re.compile(r'\.[a-zA-Z0-9]+$')  # match file extension
-        new_ext = f'.{format}'
-        durs.index = durs.index.str.replace(cur_ext, new_ext, regex=True)
-    durs.index = db.root + os.path.sep + durs.index
+        durs.index = audformat.utils.replace_file_extension(durs.index, format)
+    durs.index = audformat.utils.expand_file_path(durs.index, db.root)
     db._files_duration = durs.to_dict()