As a preparation to tackle #145, I wanted first to make caching the durations from the dependency table as fast as possible.

The changes presented here reduce the processing for loading `librispeech` from cache from 9.2 s to 8.0 s (and 5.2 s to 4.8 s on my local machine). When requesting another format (`format='wav'`) loading from cache is reduced from 24.0 s to 9.7 s.